Adding new models

Using a custom model is an advanced option outlined below. This section also serves to aid developers who want to permanently add new models to the package.

Wherever a model, such as bertalanffy or gompertz, appears in TumorGrowth.jl workflows, any function or other callable mymodel can be used, provided it has the right signature. The required signature is mymodel(times, p), where times is a vector of times and p parameters of the model in the form of a named tuple, such as p = (; v0=0.03, ω= 1.5). The return value of mymodel will be the corresponding volumes predicted by the model.

If the implementation of mymodel requires numerically solving an ordinary differential equation, follow the example given for the bertalanffy2 model, which appears here. (In the TumorGrowth.jl repository, the model ODEs themselves are defined by functions ending in _ode or _ode! in a separate file.)

Additionally, one may want to overload some of functions listed below for the new model, especially if convergence during calibration is an issue. For example, if the new model is a function is called mymodel, and there are two parameters a and b, then you overload guess_parameters like this:

function TumorGrowth.guess_parameters(times, volumes, ::typeof(mymodel))
    < code to guess parameters `a` and `b` >
	return (; a=a, b=b)
end 

Optional methods for new model implementations

TumorGrowth.guess_parameters — Function

guess_parameters(times, volumes, model)

Apply heuristics to guess parameters p for a model.

New model implementations

Fallback returns nothing which will prompt user's to explicitly specify initial parameter values in calibration problems.

TumorGrowth.scale_function — Function

scale_function(times, volumes, model)

Return an appropriate function p -> f(p) so that p = f(q) has a value of the same order of magnitude expected for model parameters, whenever q has the same form as p but with all values equal to one.

New model implementations

Fallback returns the identity.

TumorGrowth.constraint_function — Function

constraint_function(model)

Return an appropriate Bool-valued function p -> g(p) which is false whenever parameter p leaves the natural domain of model.

New model implementations

Fallback returns true always.

TumorGrowth.options — Function

options(model)

Default calibration options for model in model comparisons.

New model implementations

Fallback returns (learning_rate=0.0001, penalty=0.8)

TumorGrowth.n_iterations — Function

n_iterations(model)

Default number of iterations to run calibration of model in model comparisons.

New model implementations

Fallback returns 10000