Reference

TumorGrowth.CalibrationProblem — Method

    CalibrationProblem(times, volumes, model; learning_rate=0.0001, options...)

Specify a problem concerned with optimizing the parameters of a tumor growth model, given measured volumes and corresponding times.

Here model can be one of: bertalanffy, bertalanffy_numerical, bertalanffy2, gompertz, logistic, classical_bertalanffy.

Default optimisation is by Adam gradient descent, using a sum of squares loss. Call solve! on a problem to carry out optimisation, as shown in the example below. See "Extended Help" for advanced options, including early stopping.

Initial values of the parameters are inferred by default.

Simple solve

using TumorGrowth

times = [0.1, 6.0, 16.0, 24.0, 32.0, 39.0]
volumes = [0.00023, 8.4e-5, 6.1e-5, 4.3e-5, 4.3e-5, 4.3e-5]
problem = CalibrationProblem(times, volumes, gompertz; learning_rate=0.01)
solve!(problem, 30)    # apply 30 gradient descent updates
julia> loss(problem)   # sum of squares loss
1.7341026729860452e-9

p = solution(problem)
julia> pretty(p)
"v0=0.0002261  v∞=2.792e-5  ω=0.05731"

extended_times = vcat(times, [42.0, 46.0])
julia> gompertz(extended_times, p)[[7, 8]]
2-element Vector{Float64}:
 3.374100207406809e-5
 3.245628908921241e-5

Extended help

Solving with iteration controls

Continuing the example above, we may replace the number of iterations, n, in solve!(problem, n), with any control from IterationControl.jl:

using IterationControl
solve!(
  problem,
  Step(1),            # apply controls every 1 iteration...
  WithLossDo(),       # print loss
  Callback(problem -> print(pretty(solution(problem)))), # print parameters
  InvalidValue(),     # stop for ±Inf/NaN loss, incl. case of out-of-bound parameters
  NumberSinceBest(5), # stop when lowest loss so far was 5 steps prior
  TimeLimit(1/60),    # stop after one minute
  NumberLimit(400),   # stop after 400 steps
)
p = solution(problem)
julia> loss(problem)
7.609310030658547e-10

See IterationControl.jl for all options.

Visualizing results

using Plots
scatter(times, volumes, xlab="time", ylab="volume", label="train")
plot!(problem, label="prediction")

Models

A model can be any callable object returning volumes given times and parameters, as in predicted_volumes = model(times, p). TumorGrowth.jl supplies the following built-in models:

model	description	parameters, p	analytic?	ODE

bertalanffy	generalized Bertalanffy (GB)	(; v0, v∞, ω, λ)	yes	TumorGrowth.bertalanffy_ode

bertalanffy_numerical	generalized Bertalanffy (testing only)	(; v0, v∞, ω, λ)	no	TumorGrowth.bertalanffy_ode

bertalanffy2	2D extension of generalized Bertalanffy	(; v0, v∞, ω, λ, γ)	no	TumorGrowth.bertalanffy2_ode!

gompertz	Gompertz (GB, λ=0)	(; v0, v∞, ω)	yes	TumorGrowth.bertalanffy_ode

logistic	logistic/Verhulst (GB, λ=-1)	(; v0, v∞, ω)	yes	TumorGrowth.bertalanffy_ode

classical_bertalanffy	classical Bertalanffy (GB, λ=1/3)	(; v0, v∞, ω)	yes	TumorGrowth.bertalanffy_ode

neural2(rng, network)	neural2 ODE with Lux.jl network	(; v0, v∞, θ)	no	TumorGrowth.neural_ode

In every case v0 is the initial volume (so that predicted_volumes[1] == v0). Note this can be different from volumes[1].

Keyword options

p0: initial value of the model parameters (always a vector); inferred by default for built-in models

g=(p-> true): constraint function: If g(p) == false for some parameter p, then a warning is given and solution(problem) is frozen at the last constrained value of p; use solve!(problem, Step(1), InvalidValue(), ...) to ensure early stopping (which works because IterationControl.loss(problem) will return Inf in that case). If unspecified, the constraint function is inferred in the case of built-in models and parameters are otherwise unconstrained.

frozen: a named tuple, such as (; v0=nothing, λ=1/2); indicating parameters to be frozen at specified values during optimization; a nothing value means freeze at initial value.

learning_rate=0.0001: learning rate for Adam gradient descent optimiser

optimiser=Optimisers.Adam(learning_rate): optimiser; must be from Optimisers.jl.

scale: a scaling function with the property that p = scale(q) has a value of the same order of magnitude for the problem parameters being optimised, whenever q has the same form as p but with all values equal to one. Scaling ensures gradient descent learns all components of p at a similar rate. If unspecified, scaling is inferred for built-in models, and uniform otherwise.

half_life=Inf: set to a real positive number to replace the sum of squares loss with a weighted version; weights decay in reverse time with the specified half_life

penalty=0.0 (range=$[0, ∞)$): the larger the positive value, the more a loss function modification discourages large differences in v0 and v∞ on a log scale. Helps discourage v0 and v∞ drifting out of bounds in models whose ODE have a singularity at the origin (all built-in models except the neural network models).

ode_options...: optional keyword arguments for the ODE solver, DifferentialEquations.solve, from DifferentialEquations.jl. Not relevant for models using analytic solutions (see table above).

TumorGrowth.CalibrationProblem — Method

CalibrationProblem(problem; kwargs...)

Construct a new calibration problem out an existing problem but supply new keyword arguments, kwargs. Unspecified keyword arguments fall back to defaults, except for p0, which falls back to solution(problem).

TumorGrowth.WeightedL2Loss — Type

WeightedL2Loss(times, h=Inf)

Private method.

Return a weighted sum of squares loss function (ŷ, y) -> loss, where the weights decay in reverse time with a half life h.

TumorGrowth.bertalanffy — Method

bertalanffy(times, p)

Return volumes for specified times, based on the analytic solution to the generalized Bertalanffy model for lesion growth. Here p will have properties v0, v∞, ω, λ, where v0 is the volume at time times[1] and the other parameters are explained in the TumorGrowth.bertalanffy_ode document string.

See also bertalanffy2.

TumorGrowth.bertalanffy2 — Method

bertalanffy2(times, p; aspirational=false, solve_kwargs...)

Return volumes for specified times, based on numerical solutions to a two-dimensional extension of generalized Bertalanffy model for lesion growth. Here Here p will have properties v0, v∞, ω, λ, γ, where v0 is the volume at time times[1] and the other parameters are explained in the TumorGrowth.bertalanffy2_ode! document string.

The usual generalized Bertalanffy model is recovered when γ=0. In that case, using bertalanffy, which is based on an analytic solution, may be preferred.

Important

It is assumed without checking that times is ordered: times == sort(times).

Keyword options

aspirational=false: Set to true to return the aspirational volumes, in addition to the actual volumes.

solve_kwargs: optional keyword arguments for the ODE solver, DifferentialEquations.solve, from DifferentialEquations.jl.

See also bertalanffy.

TumorGrowth.bertalanffy2_ode! — Method

bertalanffy2_ode!(dX, X, p, t)

A two-dimensional extension of the generalized TumorGrowth.model for lesion growth (see bertalanffy_ode). Here X = [v, u], where v is volume at time t and u is the "aspirational volume" at time t, a latent variable. The time derivatives are written to dX. Specifically, dX will have these components:

$dv/dt = ω((u/v)^λ - 1)/λ)v$ $du/dt = γωu;$

where [ω, λ, γ] == p are fixed parameters:

$1/ω$ has units of time

$λ$ is dimensionless

$γ$ is dimensionless

When $γ = 0$ the model collapses to the (one-dimensional) TumorGrowth.model. In that special case, $λ = -1$, gives the logistic (Verhulst) model, while $λ = 1/3$ gives the classical TumorGrowth.model. In the case $λ = 0$, the implementation replaces $((u/v)^λ - 1)/λ$ with its limiting value $\log(u/v)$ to recover the Gompertz model when also $γ = 0$.

Since u is a latent variable, its initial value is an additional model parameter.

TumorGrowth.bertalanffy_numerical — Method

bertalanffy_numerical(times, p; solve_kwargs...)

Provided for testing purposes.

Return volumes for specified times, based on numerical solutions to the generalized Bertalanffy model for lesion growth. Here p will have properties v0, v∞, ω, λ, where v0 is the volume at time times[1] and the other parameters are explained in the TumorGrowth.bertalanffy_ode document string; solve_kwargs are optional keyword arguments for the ODE solver, DifferentialEquations.solve, from DifferentialEquations.jl.

Since it is based on analtic solutions, bertalanffy is the preferred alternative to this function.

Important

It is assumed without checking that times is ordered: times == sort(times).

See also bertalanffy2.

TumorGrowth.bertalanffy_ode — Method

bertalanffy_ode(v, p, t)

Based on the general TumorGrowth.model, return the rate in change in volume at time t, for a current volume of v. This first order, autonomous, one-dimensionaly ODE model is given by

$dv/dt = ω((v∞/v)^λ - 1)/λ)v; v>0$

where [v∞, ω, λ] == p are parameters::

$v∞$ is the steady state solution, stable and unique, assuming $ω > 0$

$1/ω$ has the units of time

$λ$ is dimensionless

When $λ = -1$, one recovers the logistic (Verhulst) model, while $λ = 1/3$ gives the classical TumorGrowth.model. In the case $λ = 0$, the implementation replaces $((v∞/v)^λ - 1)/λ$ with its limiting value $\log(v∞/v)$ to recover the Gompertz model.

TumorGrowth.classical_bertalanffy — Method

classical_bertalanffy(times, v0, v∞, ω)

Return volumes for specified times, based on anaytic solutions to the classical Bertalanffy model for lesion growth. Here p will have properties v0, v∞, ω, where v0 is the volume at time times[1] and the other parameters are explained in the TumorGrowth.bertalanffy_ode document string.

This is the λ=1/3 case of the bertalanffy model.

See also bertalanffy, bertalanffy2.

TumorGrowth.compare — Method

compare(times, volumes, models; holdouts=3, metric=mae, advanced_options...)

By calibrating models using the specified patient times and lesion volumes, compare those models using a hold-out set consisting of the last holdouts data points.

times = [0.1, 6.0, 16.0, 24.0, 32.0, 39.0]
volumes = [0.00023, 8.4e-5, 6.1e-5, 4.3e-5, 4.3e-5, 4.3e-5]

julia> comparison = compare(times, volumes, [gompertz, logistic])
ModelComparison with 3 holdouts:
  metric: mae
  gompertz:     2.198e-6
  logistic:     6.55e-6

julia> errors(comparison)
2-element Vector{Float64}:
 2.197843662660861e-6
 6.549858321487298e-6

julia> p = parameters(comparison)[1]  # calibrated parameter for `gompertz`
(v0 = 0.00022643603114569068, v∞ = 3.8453274218216947e-5, ω = 0.11537512108224635)

julia> gompertz(times, p)
6-element Vector{Float64}:
 0.00022643603114569068
 9.435316392754094e-5
 5.1039159299783234e-5
 4.303209015899451e-5
 4.021112910411027e-5
 3.922743006690166e-5

Visualing comparisons

using Plots
plot(comparison, title="A comparison of two models")

Keyword options

holdouts=3: number of time-volume pairs excluded from the end of the calibration data

metric=mae: metric applied to holdout set; the reported error on a model predicting volumes v̂ is metric(v̂, v) where v is the last holdouts values of volumes. For example, any regression measure from StatisticalMeasures.jl can be used here. The built-in fallback is mean absolute error.

n_iterations=TumorGrowth.n_iterations.(models): a vector of iteration counts for the calibration of models

calibration_options=TumorGrowth.options.(models): a vector of named tuples providing the keyword arguments for CalibrationProblems - one for each model. See CalibrationProblem for details.

See also errors, parameters.

TumorGrowth.curvature — Method

curvature(xs, ys)

Return the coefficient a for the parabola x -> a*x^2 + b*x + c of best fit, for ordinates xs and coordinates ys.

TumorGrowth.errors — Method

errors(comparison)

Extract the the vector of errors from a ModelComparison object, as returned by calls to compare.

TumorGrowth.flat_patient_data — Method

flat_patient_data()

Return, in row table form, the lesion measurement data collected in Laleh et al. (2022) "Classical mathematical models for prediction of response to chemotherapy and immunotherapy", PLOS Computational Biology.

Each row represents a single measurement of a single lesion on some day.

See also patient_data, in which each row represents all measurements of a single lesion.

TumorGrowth.gompertz — Method

gompertz(times, p)

Return volumes for specified times, based on anaytic solutions to the classical Gompertz model for lesion growth. Here p will have properties v0, v∞, ω, where v0 is the volume at time times[1] and the other parameters are explained in the TumorGrowth.bertalanffy_ode document string.

This is the λ=0 case of the bertalanffy model.

See also bertalanffy, bertalanffy2.

TumorGrowth.logistic — Method

logistic(times, v0, v∞, ω)

Return volumes for specified times, based on anaytic solutions to the classical logistic (Verhulst) model for lesion growth. Here p will have properties v0, v∞, ω, where v0 is the volume at time times[1] and the other parameters are explained in the TumorGrowth.bertalanffy_ode document string.

This is the λ=-1 case of the bertalanffy model.

See also bertalanffy, bertalanffy2.

TumorGrowth.loss — Method

loss(problem)

Return the sum of squares loss for a calibration problem, as constructed with CalibrationProblem.

TumorGrowth.neural2 — Method

neural2([rng,] network)

Initialize the Lux.jl neural2 network, network, and return a callable object, model, for solving the associated neural2 ODE for volume growth, as detailed under "The ODE" below.

Important

Here network must accept two inputs and deliver two outputs. For purposes of calibration, it may be helpful to use zero-initialization for the first layer. See the example below.

The returned object model is called like this:

volumes = model(times, p)

where p should have properties v0, v∞, θ, where v0 is the initial volume (so that first(volumes) = v0), v∞ is a volume scale parameter, and θ is a network-compatible Lux.jl parameter.

The form of θ is the same as TumorGrowth.initial_parameters(model), which is also the default initial value used when solving an associated CalibrationProblem.

using Lux, Random

# define neural2 network with 2 inputs and 2 outputs:
network = Lux.Chain(Dense(2, 3, Lux.tanh; init_weight=Lux.zeros64), Dense(3, 2))

rng = Xoshiro(123)
model = neural2(rng, network)
θ = TumorGrowth.initial_parameters(model)
times = [0.1, 6.0, 16.0, 24.0, 32.0, 39.0]
v0, v∞ = 0.00023, 0.00015
p = (; v0, v∞, θ)

julia> volumes = model(times, p) # (constant because of zero-initialization)
6-element Vector{Float64}:
 0.00023
 0.00023
 0.00023
 0.00023
 0.00023
 0.00023

The ODE

...

See also TumorGrowth.neural_ode.

TumorGrowth.neural_ode — Method

neural_ode([rng,] network)

Initialize the Lux.jl neural2 network, network, and return an associated ODE, ode, with calling syntax dX_dt = ode(X, p, t), where p is a network-compatible parameter.

The initialized parameter value can be recovered with TumorGrowth.initial_parameters(ode). Get the network state with TumorGrowth.state(ode).

using Lux
using Random

rng = Xoshiro(123)
network = network = Lux.Chain(Lux.Dense(2, 3, Lux.tanh), Lux.Dense(3, 2))
ode = neural_ode(rng, network)
θ = TumorGrowth.initial_parameters(ode)
ode(rand(2), θ, 42.9) # last argument irrelevant as `ode` is autonomous

TumorGrowth.parameters — Method

errors(comparison)

Extract the the vector of errors from a ModelComparison object, as returned by calls to compare.

TumorGrowth.patient_data — Method

patient_data()

Return, in row table form, the lesion measurement data collected in Laleh et al. (2022) "Classical mathematical models for prediction of response to chemotherapy and immunotherapy", PLOS Computational Biology.

Each row represents all measurements for a single lesion for a unique patient.

record = first(patient_data())

julia> record.Pt_hashID # patient identifier
"0218075314855e6ceacca856fcd4c737-S1"

julia> record.T_weeks # measure times, in weeks
7-element Vector{Float64}:
  0.1
  6.0
 12.0
 17.0
 23.0
 29.0
 35.0

julia> record.Lesion_normvol # all volumes measured, normalised by dataset max
7-element Vector{Float64}:
 0.000185364052636979
 0.00011229838600811
 8.4371439525252e-5
 8.4371439525252e-5
 1.05464299406565e-5
 2.89394037571615e-5
 8.4371439525252e-5

See also flat_patient_data.

TumorGrowth.recover — Method

recover(tuple, from)

Private method.

Return a new tuple by replacing any nothing values with the corresponding value in the from tuple, whenever a corresponding key exists, and otherwise not make the replacement.

julia> recover((x=1, y=nothing, z=3, w=nothing), (x=10, y=2, k=7))
(x = 1, y = 2, z = 3, w = nothing)

TumorGrowth.slope — Method

slope(xs, ys)

Return the slope of the line of least-squares best fit for ordinates xs and coordinates ys.

TumorGrowth.solution — Method

solution(problem)

Return to the solution to a CalibrationProblem. Normally applied after calling solve!(problem).

Also returns the solution to internally defined problems, as constructed with TumorGrowth.OptimisationProblem, TumorGrowth.CurveOptimisationProblem.

TumorGrowth.solve! — Method

solve!(problem, n)

Solve a calibration problem, as constructed with CalibrationProblem. The calibrated parameters are then returned by solution(problem).

solve!(problem, controls...)

Solve a calibration problem using one or more iteration controls, from the package IterationControls.jl. See the "Extended help" section of CalibrationProblem for examples.